requirements.txt
nltk>=3.8.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.2
numpy>=1.24.0
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
protobuf>=3.20.0
huggingface_hub>=0.19.0
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0

# Для GPU (раскомментируй если есть NVIDIA):
# torch --index-url https://download.pytorch.org/whl/cu118